Visual Studio update. Corrections in the code - Xml class. Fix in parsing function for bool type.

+++ SCInfor/source/XML/RapidXML/rapidxml_utils.hpp
@@ -37,7 +37,9 @@ namespace rapidxml
             
             // Determine stream size
             stream.seekg(0, ios::end);
-            size_t size = stream.tellg();
+            size_t size = (size_t)stream.tellg();
+			//oryginalny zapis w rapidxml...
+			//size_t size = stream.tellg();
             stream.seekg(0);   
             
             // load data and add terminating 0
